Non-OSS apps: mention Winaero Tweaker
It's a legit commonly recommended tool that has been out for years.
It offers both de-bloating and privacy features.
It works as advertised, installers are clean, and I never had any problem with it.

I am not affiliated with them in any way. Saw your list on HN and offering this addition as a user.